Fix a typo in the for loop iterator variable.
The `y` variable was reused inside of the nested `for` loop, which I believe to be a mistake. I assume this was just a copy/paste error.

It seems like `x` is being used for the second dimension of the array access (`....grid[y][x]`), so I updated the inner `for` loop to use `x`.