Update the `parcel` invocation sites and localhost ports.
Fixes https://github.com/ravens-engine/core/issues/10

To run `parcel` on the command line, if not installed globally, it should be issued via `npx`. The default port is also `1234` and not `8080`. While you can specify a port with an argument to the call, it is simpler to just update the docs to reflect the default.